+++ __tests__/homepage.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
-import Home from "../../src/pages/index";
+import Home from "../src/pages/index";
 import { axe } from "jest-axe";
-import { props } from "../data/homepageProps";
+import { props } from "./data/homepageProps";
 
 describe("Homepage Accessibility", () => {
   it("passes axe accessibility test", async () => {
