+++ src/components/swimlanes/swimlanes.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, within } from "@testing-library/react";
 import SwimLanes from "@/components/swimlanes/swimLanes";
-import { props } from "../data/homepageProps";
+import { props } from "../../../__tests__/data/homepageProps";
 
 describe("Swim Lanes component renders with expected props", () => {
   it("renders the first row", () => {
